requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
transformers>=4.40.0
grad-cam>=1.4.8
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
xgboost>=2.0.3
shap>=0.45.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.3.0
opencv-python-headless>=4.9.0
matplotlib>=3.8.0
groq>=0.9.0
python-dotenv>=1.0.1
scipy>=1.13.0
wikipedia-api>=0.6.0